tests/test_chain.py
import pytest
from app.generation.chain import RAGChain

def test_rag_chain_initialization():
    rag_chain = RAGChain()
    assert rag_chain is not None

def test_rag_chain_process():
    rag_chain = RAGChain()
    input_data = "Sample input for testing."
    output = rag_chain.process(input_data)
    assert isinstance(output, str)  # Assuming the output should be a string
    assert len(output) > 0  # Ensure that some output is generated

def test_rag_chain_with_empty_input():
    rag_chain = RAGChain()
    output = rag_chain.process("")
    assert output == "No input provided."  # Assuming this is the expected behavior for empty input

def test_rag_chain_with_invalid_input():
    rag_chain = RAGChain()
    output = rag_chain.process(None)
    assert output == "Invalid input."  # Assuming this is the expected behavior for invalid input