requirements.txt
fastapi
uvicorn
python-dotenv
pydantic-settings
langchain
langchain-community
langchain-experimental
langchain-groq
langchain-google-genai
langchain-core
google-genai
groq
qdrant-client
rank-bm25
langsmith
tiktoken
pypdf
httpx